+++ odooly.py
@@ -29,7 +29,7 @@ try:
 except ImportError:
     requests = None
 
-__version__ = '2.4.7'
+__version__ = '2.5.0'
 __all__ = ['Client', 'Env', 'HTTPSession', 'WebAPI', 'Service', 'Json2',
            'Printer', 'Error', 'ServerError',
            'BaseModel', 'Model', 'BaseRecord', 'Record', 'RecordList',